+++ src/export2D.h
@@ -7,8 +7,8 @@
 void ClearContents(const std::string& filePath);
 void Export2D(const std::vector<std::vector<double>> &vector2D,
         const std::string& filePath,
-        const int iteration);
+        const long long int iteration);
 
-void ExportItem(const int item, std::string& filePath);
+void ExportItem(const long long int item, std::string& filePath);
 void ExportItem(const double item, std::string& filePath);
 #endif
